Hook up websocket to signal page refresh needed
Get a websocket connection ready for any html page opened using the local
server.
For each codelab refreshed, trigger a message sent including the URL so that
client side knows if they need to refresh their page or not. They will only
refresh if the modified tutorial(s) is(are) the current viewed one.